+++ packages/app/src/stories/form/LmRhf.stories.tsx
@@ -17,7 +17,7 @@ import {
   LmSubmitButtonRhf,
   LmSwitchRhf,
 } from '../../../../form/src'
-import { LmDatepickerRhf } from '@tamagui-extras/date'
+import { LmDatepickerRhf } from '../../../../date/src'
 import React, { useState } from 'react'
 import { action } from '@storybook/addon-actions'
 import { useForm } from 'react-hook-form'