+++ packages/video/types/LmVideo.native.d.ts
@@ -1,3 +1,0 @@
-import { LmVideoProps } from "./videoTypes";
-export declare function LmVideo({ aspectRatio, src, posterSource, isLooping, resizeMode, autoplay, useNativeControls }: LmVideoProps): null;
-//# sourceMappingURL=LmVideo.native.d.ts.map