+++ packages/youtube/types/LmYoutubeEmbed.native.d.ts
@@ -1,4 +1,0 @@
-/// <reference types="react" />
-import { LmYoutubeEmbedProps } from './videoEmbedTypes';
-export declare function LmYoutubeEmbed({ youtubeId, aspectRatio, width, ...stackProps }: LmYoutubeEmbedProps): JSX.Element;
-//# sourceMappingURL=LmYoutubeEmbed.native.d.ts.map