+++ packages/app/CHANGELOG.md
@@ -1,5 +1,75 @@
 # app
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.8.4
+  - @tamagui-extras/youtube@1.8.4
+  - @tamagui-extras/video@1.9.4
+  - @tamagui-extras/core@1.8.4
+  - @tamagui-extras/date@1.9.6
+  - @tamagui-extras/file@1.8.4
+  - @tamagui-extras/form@1.12.4
+  - @tamagui-extras/link@1.8.4
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.8.3
+  - @tamagui-extras/youtube@1.8.3
+  - @tamagui-extras/video@1.9.3
+  - @tamagui-extras/core@1.8.3
+  - @tamagui-extras/date@1.9.5
+  - @tamagui-extras/file@1.8.3
+  - @tamagui-extras/form@1.12.3
+  - @tamagui-extras/link@1.8.3
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.8.2
+  - @tamagui-extras/youtube@1.8.2
+  - @tamagui-extras/video@1.9.2
+  - @tamagui-extras/core@1.8.2
+  - @tamagui-extras/date@1.9.4
+  - @tamagui-extras/file@1.8.2
+  - @tamagui-extras/form@1.12.2
+  - @tamagui-extras/link@1.8.2
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.8.1
+  - @tamagui-extras/youtube@1.8.1
+  - @tamagui-extras/video@1.9.1
+  - @tamagui-extras/core@1.8.1
+  - @tamagui-extras/date@1.9.3
+  - @tamagui-extras/file@1.8.1
+  - @tamagui-extras/form@1.12.1
+  - @tamagui-extras/link@1.8.1
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.8.0
+  - @tamagui-extras/youtube@1.8.0
+  - @tamagui-extras/video@1.9.0
+  - @tamagui-extras/core@1.8.0
+  - @tamagui-extras/file@1.8.0
+  - @tamagui-extras/form@1.12.0
+  - @tamagui-extras/link@1.8.0
+  - @tamagui-extras/date@1.9.2
+
 ## 0.0.26
 
 ### Patch Changes